Squashed commit of the following:
commit 90f3d2f1034abf7fede6e214a4912dbb8f4efc4f
Author: Frédéric Metrich <45318189+FredM67@users.noreply.github.com>
Date:   Thu Dec 26 18:40:11 2024 +0100

    Update copyright year to 2024 and refactor type traits for consistency

+++ Mk2_3phase_RFdatalog_temp/type_traits/is_enum.hpp
@@ -1,5 +1,5 @@
 // ArduinoJson - https://arduinojson.org
-// Copyright © 2014-2023, Benoit BLANCHON
+// Copyright © 2014-2024, Benoit BLANCHON
 // MIT License
 
 #pragma once
@@ -11,7 +11,8 @@
 #include "is_same.hpp"
 
 template< typename T >
-struct is_enum
-{
-  static const bool value = is_convertible< T, int >::value && !is_class< T >::value && !is_integral< T >::value && !is_floating_point< T >::value;
+struct is_enum {
+  static const bool value = is_convertible<T, long long>::value &&
+                            !is_class<T>::value && !is_integral<T>::value &&
+                            !is_floating_point<T>::value;
 };
